+++ packages/eslint-plugin/src/rules/camelcase.ts
@@ -8,6 +8,19 @@ import * as util from '../util';
 type Options = util.InferOptionsTypeFromRule<typeof baseRule>;
 type MessageIds = util.InferMessageIdsTypeFromRule<typeof baseRule>;
 
+const schema = util.deepMerge(
+  Array.isArray(baseRule.meta.schema)
+    ? baseRule.meta.schema[0]
+    : baseRule.meta.schema,
+  {
+    properties: {
+      genericType: {
+        enum: ['always', 'never'],
+      },
+    },
+  },
+);
+
 export default util.createRule<Options, MessageIds>({
   name: 'camelcase',
   meta: {
@@ -17,7 +30,7 @@ export default util.createRule<Options, MessageIds>({
       category: 'Stylistic Issues',
       recommended: 'error',
     },
-    schema: baseRule.meta.schema,
+    schema: [schema],
     messages: baseRule.meta.messages,
   },
   defaultOptions: [
@@ -25,6 +38,7 @@ export default util.createRule<Options, MessageIds>({
       allow: ['^UNSAFE_'],
       ignoreDestructuring: false,
       properties: 'never',
+      genericType: 'never',
     },
   ],
   create(context, [options]) {
@@ -36,11 +50,13 @@ export default util.createRule<Options, MessageIds>({
       AST_NODE_TYPES.TSAbstractClassProperty,
     ];
 
+    const genericType = options.genericType;
     const properties = options.properties;
-    const allow = (options.allow || []).map(entry => ({
-      name: entry,
-      regex: new RegExp(entry),
-    }));
+    const allow =
+      options.allow?.map(entry => ({
+        name: entry,
+        regex: new RegExp(entry),
+      })) ?? [];
 
     /**
      * Checks if a string contains an underscore and isn't all upper-case
@@ -72,21 +88,26 @@ export default util.createRule<Options, MessageIds>({
      * @private
      */
     function isTSPropertyType(node: TSESTree.Node): boolean {
-      if (!node.parent) {
-        return false;
-      }
-      if (TS_PROPERTY_TYPES.includes(node.parent.type)) {
+      if (TS_PROPERTY_TYPES.includes(node.type)) {
         return true;
       }
 
-      if (node.parent.type === AST_NODE_TYPES.AssignmentPattern) {
+      if (node.type === AST_NODE_TYPES.AssignmentPattern) {
         return (
-          node.parent.parent !== undefined &&
-          TS_PROPERTY_TYPES.includes(node.parent.parent.type)
+          node.parent !== undefined &&
+          TS_PROPERTY_TYPES.includes(node.parent.type)
         );
       }
 
       return false;
+    }
+
+    function report(node: TSESTree.Identifier): void {
+      context.report({
+        node,
+        messageId: 'notCamelCase',
+        data: { name: node.name },
+      });
     }
 
     return {
@@ -103,13 +124,32 @@ export default util.createRule<Options, MessageIds>({
         }
 
         // Check TypeScript specific nodes
-        if (isTSPropertyType(node)) {
+        const parent = node.parent;
+        if (parent && isTSPropertyType(parent)) {
           if (properties === 'always' && isUnderscored(name)) {
-            context.report({
-              node,
-              messageId: 'notCamelCase',
-              data: { name: node.name },
-            });
+            report(node);
+          }
+
+          return;
+        }
+
+        if (parent && parent.type === AST_NODE_TYPES.TSTypeParameter) {
+          if (genericType === 'always' && isUnderscored(name)) {
+            report(node);
+          }
+
+          return;
+        }
+
+        if (parent && parent.type === AST_NODE_TYPES.OptionalMemberExpression) {
+          // Report underscored object names
+          if (
+            properties === 'always' &&
+            parent.object.type === AST_NODE_TYPES.Identifier &&
+            parent.object.name === node.name &&
+            isUnderscored(name)
+          ) {
+            report(node);
           }
 
           return;