+++ packages/eslint-plugin/src/rules/no-unnecessary-condition.ts
@@ -31,6 +31,17 @@ const isPossiblyFalsy = (type: ts.Type): boolean =>
 const isPossiblyTruthy = (type: ts.Type): boolean =>
   unionTypeParts(type).some(type => !isFalsyType(type));
 
+// Nullish utilities
+const nullishFlag = ts.TypeFlags.Undefined | ts.TypeFlags.Null;
+const isNullishType = (type: ts.Type): boolean =>
+  isTypeFlagSet(type, nullishFlag);
+
+const isPossiblyNullish = (type: ts.Type): boolean =>
+  unionTypeParts(type).some(isNullishType);
+
+const isAlwaysNullish = (type: ts.Type): boolean =>
+  unionTypeParts(type).every(isNullishType);
+
 // isLiteralType only covers numbers and strings, this is a more exhaustive check.
 const isLiteral = (type: ts.Type): boolean =>
   isBooleanLiteralType(type, true) ||
@@ -41,15 +52,9 @@ const isLiteral = (type: ts.Type): boolean =>
   isLiteralType(type);
 // #endregion
 
-type ExpressionWithTest =
-  | TSESTree.ConditionalExpression
-  | TSESTree.DoWhileStatement
-  | TSESTree.ForStatement
-  | TSESTree.IfStatement
-  | TSESTree.WhileStatement;
-
 export type Options = [
   {
+    allowConstantLoopConditions?: boolean;
     ignoreRhs?: boolean;
   },
 ];
@@ -57,6 +62,8 @@ export type Options = [
 export type MessageId =
   | 'alwaysTruthy'
   | 'alwaysFalsy'
+  | 'neverNullish'
+  | 'alwaysNullish'
   | 'literalBooleanExpression'
   | 'never';
 export default createRule<Options, MessageId>({
@@ -74,6 +81,9 @@ export default createRule<Options, MessageId>({
       {
         type: 'object',
         properties: {
+          allowConstantLoopConditions: {
+            type: 'boolean',
+          },
           ignoreRhs: {
             type: 'boolean',
           },
@@ -84,6 +94,10 @@ export default createRule<Options, MessageId>({
     messages: {
       alwaysTruthy: 'Unnecessary conditional, value is always truthy.',
       alwaysFalsy: 'Unnecessary conditional, value is always falsy.',
+      neverNullish:
+        'Unnecessary conditional, expected left-hand side of `??` operator to be possibly null or undefined.',
+      alwaysNullish:
+        'Unnecessary conditional, left-hand side of `??` operator is always `null` or `undefined`',
       literalBooleanExpression:
         'Unnecessary conditional, both sides of the expression are literal values',
       never: 'Unnecessary conditional, value is `never`',
@@ -91,10 +105,11 @@ export default createRule<Options, MessageId>({
   },
   defaultOptions: [
     {
+      allowConstantLoopConditions: false,
       ignoreRhs: false,
     },
   ],
-  create(context, [{ ignoreRhs }]) {
+  create(context, [{ allowConstantLoopConditions, ignoreRhs }]) {
     const service = getParserServices(context);
     const checker = service.program.getTypeChecker();
 
@@ -110,16 +125,47 @@ export default createRule<Options, MessageId>({
     function checkNode(node: TSESTree.Node): void {
       const type = getNodeType(node);
 
+      // Conditional is always necessary if it involves:
+      //    `any` or `unknown` or a naked type parameter
+      if (
+        unionTypeParts(type).some(part =>
+          isTypeFlagSet(
+            part,
+            TypeFlags.Any | TypeFlags.Unknown | ts.TypeFlags.TypeParameter,
+          ),
+        )
+      ) {
+        return;
+      }
+      const messageId = isTypeFlagSet(type, TypeFlags.Never)
+        ? 'never'
+        : !isPossiblyTruthy(type)
+        ? 'alwaysFalsy'
+        : !isPossiblyFalsy(type)
+        ? 'alwaysTruthy'
+        : undefined;
+
+      if (messageId) {
+        context.report({ node, messageId });
+      }
+    }
+
+    function checkNodeForNullish(node: TSESTree.Node): void {
+      const type = getNodeType(node);
       // Conditional is always necessary if it involves `any` or `unknown`
       if (isTypeFlagSet(type, TypeFlags.Any | TypeFlags.Unknown)) {
         return;
       }
-      if (isTypeFlagSet(type, TypeFlags.Never)) {
-        context.report({ node, messageId: 'never' });
-      } else if (!isPossiblyTruthy(type)) {
-        context.report({ node, messageId: 'alwaysFalsy' });
-      } else if (!isPossiblyFalsy(type)) {
-        context.report({ node, messageId: 'alwaysTruthy' });
+      const messageId = isTypeFlagSet(type, TypeFlags.Never)
+        ? 'never'
+        : !isPossiblyNullish(type)
+        ? 'neverNullish'
+        : isAlwaysNullish(type)
+        ? 'alwaysNullish'
+        : undefined;
+
+      if (messageId) {
+        context.report({ node, messageId });
       }
     }
 
@@ -157,14 +203,13 @@ export default createRule<Options, MessageId>({
      * Filters all LogicalExpressions to prevent some duplicate reports.
      */
     function checkIfTestExpressionIsNecessaryConditional(
-      node: ExpressionWithTest,
+      node: TSESTree.ConditionalExpression | TSESTree.IfStatement,
     ): void {
-      if (
-        node.test !== null &&
-        node.test.type !== AST_NODE_TYPES.LogicalExpression
-      ) {
-        checkNode(node.test);
+      if (node.test.type === AST_NODE_TYPES.LogicalExpression) {
+        return;
       }
+
+      checkNode(node.test);
     }
 
     /**
@@ -173,20 +218,56 @@ export default createRule<Options, MessageId>({
     function checkLogicalExpressionForUnnecessaryConditionals(
       node: TSESTree.LogicalExpression,
     ): void {
+      if (node.operator === '??') {
+        checkNodeForNullish(node.left);
+        return;
+      }
       checkNode(node.left);
       if (!ignoreRhs) {
         checkNode(node.right);
       }
     }
 
+    /**
+     * Checks that a testable expression of a loop is necessarily conditional, reports otherwise.
+     */
+    function checkIfLoopIsNecessaryConditional(
+      node:
+        | TSESTree.DoWhileStatement
+        | TSESTree.ForStatement
+        | TSESTree.WhileStatement,
+    ): void {
+      if (
+        node.test === null ||
+        node.test.type === AST_NODE_TYPES.LogicalExpression
+      ) {
+        return;
+      }
+
+      /**
+       * Allow:
+       *   while (true) {}
+       *   for (;true;) {}
+       *   do {} while (true)
+       */
+      if (
+        allowConstantLoopConditions &&
+        isBooleanLiteralType(getNodeType(node.test), true)
+      ) {
+        return;
+      }
+
+      checkNode(node.test);
+    }
+
     return {
       BinaryExpression: checkIfBinaryExpressionIsNecessaryConditional,
       ConditionalExpression: checkIfTestExpressionIsNecessaryConditional,
-      DoWhileStatement: checkIfTestExpressionIsNecessaryConditional,
-      ForStatement: checkIfTestExpressionIsNecessaryConditional,
+      DoWhileStatement: checkIfLoopIsNecessaryConditional,
+      ForStatement: checkIfLoopIsNecessaryConditional,
       IfStatement: checkIfTestExpressionIsNecessaryConditional,
-      WhileStatement: checkIfTestExpressionIsNecessaryConditional,
       LogicalExpression: checkLogicalExpressionForUnnecessaryConditionals,
+      WhileStatement: checkIfLoopIsNecessaryConditional,
     };
   },
 });