+++ packages/eslint-plugin/tests/rules/no-unnecessary-condition.test.ts
@@ -72,13 +72,39 @@ const t1 = (b1 && b2) ? 'yes' : 'no'`,
 function test<T extends string>(t: T) {
   return t ? 'yes' : 'no'
 }`,
+    `
+// Naked type param
+function test<T>(t: T) {
+  return t ? 'yes' : 'no'
+}`,
+    `
+// Naked type param in union
+function test<T>(t: T | []) {
+  return t ? 'yes' : 'no'
+}`,
 
     // Boolean expressions
     `
 function test(a: string) {
   return a === "a"
 }`,
-
+    // Nullish coalescing operator
+    `
+function test(a: string | null) {
+  return a ?? "default";
+}`,
+    `
+function test(a: string | undefined) {
+  return a ?? "default";
+}`,
+    `
+function test(a: string | null | undefined) {
+  return a ?? "default";
+}`,
+    `
+function test(a: unknown) {
+  return a ?? "default";
+}`,
     // Supports ignoring the RHS
     {
       code: `
@@ -86,6 +112,14 @@ declare const b1: boolean;
 declare const b2: true;
 if(b1 && b2) {}`,
       options: [{ ignoreRhs: true }],
+    },
+    {
+      code: `
+while(true) {}
+for (;true;) {}
+do {} while(true)
+      `,
+      options: [{ allowConstantLoopConditions: true }],
     },
   ],
   invalid: [
@@ -169,6 +203,35 @@ if (x === Foo.a) {}
 `,
       errors: [ruleError(8, 5, 'literalBooleanExpression')],
     },
+    // Nullish coalescing operator
+    {
+      code: `
+function test(a: string) {
+  return a ?? 'default';
+}`,
+      errors: [ruleError(3, 10, 'neverNullish')],
+    },
+    {
+      code: `
+function test(a: string | false) {
+  return a ?? 'default';
+}`,
+      errors: [ruleError(3, 10, 'neverNullish')],
+    },
+    {
+      code: `
+function test(a: null) {
+  return a ?? 'default';
+}`,
+      errors: [ruleError(3, 10, 'alwaysNullish')],
+    },
+    {
+      code: `
+function test(a: never) {
+  return a ?? 'default';
+}`,
+      errors: [ruleError(3, 10, 'never')],
+    },
 
     // Still errors on in the expected locations when ignoring RHS
     {
@@ -189,6 +252,19 @@ const t1 = (b1 && b2) ? 'yes' : 'no'`,
         ruleError(7, 7, 'alwaysTruthy'),
         ruleError(8, 18, 'alwaysTruthy'),
         ruleError(9, 13, 'alwaysTruthy'),
+      ],
+    },
+    {
+      code: `
+while(true) {}
+for (;true;) {}
+do {} while(true)
+      `,
+      options: [{ allowConstantLoopConditions: false }],
+      errors: [
+        ruleError(2, 7, 'alwaysTruthy'),
+        ruleError(3, 7, 'alwaysTruthy'),
+        ruleError(4, 13, 'alwaysTruthy'),
       ],
     },
   ],