+++ packages/eslint-plugin/tests/rules/no-unnecessary-type-arguments.test.ts
@@ -18,6 +18,10 @@ ruleTester.run('no-unnecessary-type-arguments', rule, {
       f();`,
     `function f<T = number>() { }
       f<string>();`,
+    `declare const f: (<T = number>() => void) | null;
+      f?.();`,
+    `declare const f: (<T = number>() => void) | null;
+      f?.<string>();`,
     `declare const f: any;
       f();`,
     `declare const f: any;
@@ -53,6 +57,9 @@ ruleTester.run('no-unnecessary-type-arguments', rule, {
     `declare const C: unknown;
       class D<TD = number> extends C { }`,
     `let a: A<number>`,
+    `class Foo<T> {}
+      const foo = new Foo<number>();`,
+    `type Foo<T> = import('foo').Foo<T>;`,
   ],
   invalid: [
     {
@@ -122,6 +129,17 @@ ruleTester.run('no-unnecessary-type-arguments', rule, {
       ],
       output: `interface I<T = number> { }
         class Impl implements I { }`,
+    },
+    {
+      code: `class Foo<T = number> {}
+        const foo = new Foo<number>();`,
+      errors: [
+        {
+          messageId: 'unnecessaryTypeParameter',
+        },
+      ],
+      output: `class Foo<T = number> {}
+        const foo = new Foo();`,
     },
   ],
 });